Scripts and feature files are added
Matab scripts to generate the features for the raw sensor data have been
stored in \MoST\Matlab\featureExtractionScripts.
Procedure to run the scripts is explained in a readme.txt file in the
same folder.
The new feature files are stored in the respective raw data folders.
The features extracted are listed in the script extractSensorFeatures.m